Put error class in own file
